requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-cors-headers==3.5.0
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-gis==0.16
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
